requirements.txt
asgiref==3.8.1
beautifulsoup4==4.12.3
certifi==2024.8.30
charset-normalizer==3.3.2
Django==5.1.1
djangorestframework==3.15.2
frozendict==2.4.4
html5lib==1.1
idna==3.10
lxml==5.3.0
multitasking==0.0.11
numpy==2.1.1
pandas==2.2.3
peewee==3.17.6
platformdirs==4.3.6
psycopg2-binary==2.9.9
python-dateutil==2.9.0.post0
pytz==2024.2
requests==2.32.3
six==1.16.0
soupsieve==2.6
sqlparse==0.5.1
tzdata==2024.2
urllib3==2.2.3
webencodings==0.5.1
yfinance==0.2.43
